Move profile model to dachor_internal app
